move type filter to model scope
